Refactor: extract DonationAddressValidation
Move donation address validation to its own class. Will be needed for
atomic transaction validation.
