+++ core/src/main/java/bisq/core/trade/protocol/tasks/buyer/BuyerVerifiesPreparedDelayedPayoutTx.java
@@ -18,6 +18,7 @@
 package bisq.core.trade.protocol.tasks.buyer;
 
 import bisq.core.trade.DelayedPayoutTxValidation;
+import bisq.core.trade.DonationAddressValidation;
 import bisq.core.trade.Trade;
 import bisq.core.trade.protocol.tasks.TradeTask;
 
@@ -43,7 +44,7 @@ public class BuyerVerifiesPreparedDelayedPayoutTx extends TradeTask {
                     processModel.getBtcWalletService());
 
             complete();
-        } catch (DelayedPayoutTxValidation.DonationAddressException |
+        } catch (DonationAddressValidation.DonationAddressException |
                 DelayedPayoutTxValidation.MissingDelayedPayoutTxException |
                 DelayedPayoutTxValidation.InvalidTxException |
                 DelayedPayoutTxValidation.InvalidLockTimeException |
